Fix multiple-override race for architecture-independent binaries.
If an architecture-independent binary had multiple identical overrides
created in the same publisher run (for example by multiple archive
administrators racing to fulfil the same request, or by the
phased-updates script), then BinaryPackagePublishingHistory.supersede
would supersede them all without ensuring that the newest publication
remained live.

The architecture-independent sibling publications don't have a formal
relation to each other in the data model, which is why this has always
been hard to fix.  However, the dominator sorts publications with equal
versions by their creation dates, so it will always consider the most
recent override to be live.  We can therefore fix this by deferring the
superseding of architecture-independent siblings until we've positively
confirmed which publications are live using the dominator's normal
logic, and taking care not to supersede the confirmed-live publications.

An additional complication is that the dominator processes one
architecture at a time, so in general it won't have confirmed the full
set of live publications until it has processed all architectures.  To
cope with this, I've lifted the handling of architecture-independent
siblings out of BinaryPackagePublishingHistory.supersede into a new
BinaryPackagePublishingHistory.supersedeAssociated method.  The
supersede method calls supersedeAssociated by default to minimise chaos
in the test suite, but the dominator passes a flag to disable that and
calls supersedeAssociated separately.