Support encrypting snap store discharge macaroons at rest.
We should be encrypting sensitive data such as authentication tokens at rest
in the database, and this adds basic general support for doing that and sets
it up for the SSO discharge macaroons used for uploading snaps to the store.

These discharge macaroons are refreshed every so often, and that will
encrypt the refreshed tokens if a suitable public key is configured.

The JSON-serialised encrypted format includes the public key in order to
support future key rotation (by configuring multiple key pairs and using the
corresponding private key for decryption), although I haven't actually
written the code for that yet.