Improve caching of several delegated authorization checks.
forwardCheckAuthenticated and forwardCheckUnauthenticated don't
currently have access to the cache used by iter_authorization when doing
delegated authorization checks.  In several cases, this is the only
thing called by checkAuthenticated/checkUnauthenticated, and in that
situation a simpler syntax is available: we can just yield the object
and permission to check, and iter_authorization will do the delegated
check for us.  This makes a significant difference when operating on
many publishing objects in a single private archive: in the
archive.getPublishedBinaries webservice method, about two-thirds of the
non-trivial archive visibility checks that were previously uncached are
now cached.

I think it's possible to do better, but it would require giving security
adapters access to iter_authorization's cache so that forwardCheck* can
use it when doing delegated checks, which is a somewhat more involved
change.