Include metadata-only bug changes in Person:+commentedbugs.
Some kinds of changes that users can make to bugs are currently
invisible to all bug searches: if a user just changes some bug metadata
(title, status, etc.) then that can't be searched for, because it isn't
internally represented as a comment.  On the other hand, bug web pages
render metadata changes in a way that looks somewhat similar to
comments, so this is a bit confusing.  It's also a problem when dealing
with certain kinds of abuse.

I think the simplest option is to just include these kinds of changes
under "Commented bugs" (and hence also under "Related bugs").  There may
be some cases where that's a little surprising, but I think they'll be
outweighed by the added usefulness.