Use timeout(1) to limit debdiff rather than using alarm(3) ourselves.
The approach of calling alarm(3) and then execing the target command
relies on the target command not changing the disposition of SIGALRM,
and may have other problems (we've observed PackageDiffJobs occasionally
hanging indefinitely, although we don't know exactly why that's
happening).  coreutils has a perfectly good timeout(1) program that we
can use instead, which runs the target command as a child process and in
general seems likely to be more robust.