Add optional geoip2 support.
We don't have a useful way to test this until the relevant data packages
gain the right set of database files, but I've run the existing tests with
modified configuration to point to a local download of GeoLite2-Country.mmdb
and they pass.

Once we're using this, it will give us GeoIP support for IPv6 addresses.