Refactor the way macaroon verification handles users.
To minimise the chance of accidents, macaroon verification now records the
user for whom the macaroon was verified, or NO_USER if the macaroon was
positively verified as not having been issued on behalf of a particular
user.  The Git XML-RPC API and the authserver double-check that the recorded
user matches what they expect, thereby avoiding problems due to a particular
issuer implementation forgetting to perform user verification.

SnapBuildMacaroonIssuer and BinaryPackageBuildMacaroonIssuer now refuse
verification if given a user.