Remove parts of lp.code.mail.patches that are unmodified from bzrlib.
lp.code.mail.patches was cloned-and-hacked from bzrlib.patches in order to
support git diffs, but at the time we were in a rush and ended up copying
the whole module rather than just the parts we needed to change.  This
removes all the parts that were identical except for trivial details like
whitespace.

This will simplify porting to breezy.