Fix GitRepositorySet.new, and export it on the webservice.
This was previously untested, and had been broken since r18222.

Now that namespace.createRepository can create the repository on the hosting
service, we can safely export this to create a bare repository via the API.
For example, this allows snapcraft to create a temporary repository, issue
an access token to it, and push code to it, all without needing to configure
an SSH key.

LP: #1824399