Remove tickcount and track traversal/publication durations instead.
tickcount can't work on Python 3, so in the long term we don't have much
choice here.  We do want to be able to spot contention, but tracking thread
time should be enough here.  (Unfortunately, as far as I can tell this is
only possible with Python >= 3.3.)

This does change the access log format slightly (logging durations rather
than ticks), but as far as I can tell nothing cares.