Convert lp.testing to print_function
I also added unicode_literals in a few places where it could be done
with minimal disruption.

The build_test_suite change is because for doctests we can't just do
"from __future__ import print_function" etc. at the top of the file, and
instead have to insert print_function etc. into their globals when
constructing the corresponding test suite; this is a way of spelling
that without having to incur too much verbiage at every site.