Rename OCIRecipeTarget to OCIProject
Other *Target names in Launchpad are interfaces rather than structural
elements, so OCIRecipeTarget wasn't a great choice.

We haven't been able to think of a perfect name, but OCIProject isn't
too bad: it conveys the notion of this being a container, and this does
share some characteristics with (sub)projects.  The only major downside
is that it's going to look a little odd when we have OCIProjects within
projects (in the pillar sense), but that's probably tolerable.

Similarly, rename OCIRecipeName to OCIProjectName, and
OCIRecipeTargetSeries to OCIProjectSeries.