Replace TestEventListener with ZopeEventHandlerFixture
After adding optional support to ZopeEventHandlerFixture for specifying
required interfaces, it becomes a superset of the old TestEventListener,
with two clear advantages: it can unregister event handlers on cleanup,
and it uses the fixture pattern which is much more common for this sort
of thing nowadays.  Convert all users of TestEventListener to
ZopeEventHandlerFixture.

(This is very slightly more verbose in doctests because we have to call
the fixture's setUp method separately; but this is minor, and in
exchange self.useFixture makes it less verbose in tests using
testtools.)