Remove all remaining uses of zope.app.testing
The main work here is in refactoring functional test setup to use the
new style based on zope.component.testlayer.ZCMLFileLayer and
zope.testbrowser.wsgi.Layer.  What used to be in wsgi_application is now
pulled apart into a few pieces of middleware around
zope.app.wsgi.WSGIPublisherApplication.

The extent to which we have to pretend that ZODB exists during test
setup is now even smaller: we just need a couple of adjustments in
lp.testing.layers._FunctionalBrowserLayer.

The zcml directory now doubles as a (trivial) package, since
ZCMLFileLayer expects that.

We have to be a bit more careful in a couple of places:

 * PageTestLayer now removes its log handler on tearDown.
 * LaunchpadWebServiceCaller.addHeadersTo makes sure that the headers it
   adds are encoded in WSGI-friendly ways.
 * A couple of tests in TestBuilderHistoryView need to clear the
   permission cache between creating test objects and doing
   privacy-related tests on them; I haven't found exactly what was doing
   this before, but I think it may have been an accidental side-effect
   of something in the guts of ZODB.