Refactor lp.testing.browser using zope.testbrowser.wsgi
This test browser instance is used for a few tests that use
AppServerLayer and make out-of-process connections to the test app
server via a test browser.

The old mechanize-based test browser is going away, so we need to do
something.  A reasonably future-proof approach seems to be to use the
new WSGI test browser and fool it into connecting to the app server
using WSGIProxy2 (which in fact WebTest also uses in some situations).
We'll still need some slight adjustments here to upgrade to newer
versions of zope.testbrowser, but they should be much more manageable.

We do need to take some special care in a few places:

 * TestOpenIDReplayAttack can't pass in a custom mech_browser any more.
   For now, we switch off redirect handling entirely and follow
   redirections manually; this will need further rearrangements for
   zope.testbrowser >= 5.0.0.

 * Both the httplib and the requests clients offered by WSGIProxy2
   incorrectly fold multiple Set-Cookie headers in the response into
   one, joining them with commas; this causes test failures.
   Fortunately, the urllib3 client doesn't have this flaw, but we do
   need to take care to disable certificate verification.

 * The new WSGI test browser needs to be monkey-patched to allow talking
   to our various test hosts, as by default it only allows localhost,
   127.0.0.1, *.example.com, *.example.net, and *.example.org.