Fix unbalanced tags in ProductBugTrackerWidget
If the project is part of a project group that has a bug tracker, then
there was a superfluous "</label>".  Amazingly, this has been wrong
since 2006 and nobody ever seems to have noticed; but it confuses
zope.testbrowser 5.x.

Since the new zope.testbrowser will catch this once we upgrade to it, I
didn't think it was worth adding a separate test.