Add an API for Mailman to update team addresses
The only part of Launchpad's Mailman integration that requires the same
deployment to have access to both Mailman and the Launchpad database is
the mlist-sync script, which runs on staging and needs to update the
team email addresses corresponding to each mailing list to refer to the
staging hostname rather than the production hostname.  Decoupling this
will allow extracting Launchpad's customised version of Mailman from the
Launchpad tree and running it separately.

The simplest approach seems to be a single endpoint that just updates
all the team mailing list addresses in bulk.  For safety, this new
endpoint refuses to do anything if it finds that it's running on a
non-demo site.