Stop using launchpadlib in lp.blueprints.tests.test_webservice
Port the blueprints webservice tests to use in-process webservice calls
rather than launchpadlib and AppServerLayer.  While the code is a bit
longer as a result, it's easier to debug and substantially faster: this
change takes the test time for this file from 149 seconds to 63 seconds
on my laptop.  It also removes the last remaining dependency on
AppServerLayer's SMTP server outside of the Mailman integration tests.