Add some slack to IPoll.saneDates invariant
The IPoll.saneDates invariant required that a poll's opening time must
be no less than 12 hours after it's created.  However, the creation time
is computed by the invariant, while the opening time is supplied in the
creation form, thus causing obvious problems in boundary cases depending
on how long it takes to get as far as checking the invariant.
lib/lp/registry/browser/tests/poll-views_0.txt seems to hit this
consistently after upgrading ZTK packages to modern versions; I didn't
investigate exactly what caused this since the logical problem is
reasonably clear.

The twelve-hour limit seems pretty arbitrary, and it's not clear how to
fix this "correctly"; but adding an extra minute of slack to the
invariant does no harm and fixes the test failure.