Abort transaction in Celery after-commit hook
With transaction >= 1.6.0, Celery-related tests started failing in a
very confusing way.  After a construct such as this, which is common in
tests:

    with block_on_job():
        transaction.commit()

... subsequent test code would act as though the job hadn't been run,
even though investigation showed that it clearly had.

The proximate cause was this change in transaction 1.6.0:

    - Drop references to data managers joined to a transaction when it is
      committed or aborted.

Crucially, this happens before running after-commit hooks.  As a result,
Storm's data manager no longer arranged to implicitly close the
transaction that was implicitly opened by querying job details in
BaseJobRunner.runViaCelery.  This meant that subsequent code would see a
snapshot of the database taken before the job was run, even though it
appeared at a glance as though anything textually following the
`transaction.commit()` above should be in a new transaction (when not
thinking carefully about the order of execution of context managers).

It would be possible to work around this in each test, but it's really
far too confusing to be sensible.  Fortunately, there's a simple fix:
since runViaCelery never issues any data-modifying statements itself,
celeryCommitHook can just roll back the transaction after running it,
ensuring that subsequent statements will run in a new transaction.