Fix py-amqp deprecation warning
py-amqp was issuing the deprecation warning "The auto_delete flag for
exchanges has been deprecated and will be removed from py-amqp v1.5.0".
Use Fixture.addCleanup to delete the exchange instead.

The auto_delete flag for queues wasn't deprecated, but it seems to make
sense to handle it in the same way, since this way the CaptureOops
fixture is in control of both lifetimes.