Add OCIRecipe.build_daily column
It seems very likely that OCI images will include things that can't
easily be made to fire webhooks when they're updated, such as the
contents of random apt archives (or indeed of Ubuntu itself).  As such
we're likely to want some kind of periodic build support.

This is undeniably crude: it only supports daily builds or manual
requests.  However, it's at least simple, and it seems unlikely that
we're going to want to spend time building complex period controls for
the time being.  The column name matches the one in SourcePackageRecipe.