Export build.source_package_version
It's remarkably hard to figure out the source package version for a
binary package publication via the API.  You can sometimes do it via
`bpph.build.getLatestSourcePublication()`, but that might not be visible
if the build was performed in a private archive; or you can look at
`bpph.build.title`, but that involves parsing non-machine-readable text.

`build.source_package_name` is already exported, so it seems to make
sense to export `build.source_package_version` to go with it.