Copy html5browser into our tree
This is from lp:html5-browser, with only import path changes.

We use this for YUI tests, but the Launchpad and PyPI projects for it
have a single maintainer who doesn't seem responsive, and we need to fix
some deprecation warnings and do some other maintenance.  Fortunately
it's a thin wrapper around WebKit, so the simplest approach is to just
take our own copy.