Cache Person.name for attach_http_request
Commit e6db69425adec2ce77d67d210b6bf1cead298bb9 introduced a regression:
if the request timed out, then the attempt to fetch
principal.person.name in attach_http_request could itself raise
RequestExpired if the relevant property wasn't already cached, causing
the OOPS to be lost.

Instead, we cache Person.name on LaunchpadPrincipal earlier in the
request, allowing us to use it safely while building an OOPS report.