Port xx-announcements.txt too
It uses lp.services.feeds.tests.helper, so needs to be ported along with
the other feed tests.

By default, Beautiful Soup 4 produces output with HTML entities
converted to Unicode characters.  I could have used formatter="html" or
similar to restore something closer to the old behaviour in this case,
but the new behaviour is easier to read.