Remove use of urllib.splittype
Although it's still present in the Python 3 standard library, it's no
longer in __all__, and it isn't supported by six.moves.urllib.parse
which makes it awkward to port.  Use urlsplit instead.

This may slightly change the behaviour of
bugtracker.base_url_permutations in some cases of invalid URL input, but
I don't think that's a big deal.