Send prerequisite repository path for merge diffs
When generating preview diffs for Git-based merge proposals with
prerequisites, we send the commit IDs of the target, source, and
prerequisite to the hosting service, but the combined path that we sent
only included the target and source repository paths.  As a result, if
the tip of the prerequisite branch wasn't in either the target or the
source repository (perhaps because the prerequisite branch changed after
the source branch was created), the hosting service would be unable to
find that commit.

To fix this, simply add the prerequisite repository path to the combined
path.

LP: #1861095