Fix base_url_permutations following removal of urllib.splittype
It turns out that one test did depend on the behaviour of
base_url_permutations when given a particular kind of invalid URL, that
is, one with neither scheme nor netloc (e.g. what you get if you just
omit http:// or https:// from the start of a normal HTTP(S) URL).  This
seems like a reasonable thing to support, so adjust things slightly to
do so.