Explicitly depend on futures
We've been depending on gunicorn[gthread] and expecting that to pull in
futures.  Unfortunately this gets confusing when trying to build for
Python 3.  Rather than depending on 'futures; python_version <= "3.2"',
gunicorn adds the dependency in setup.py depending on whether it's
running on Python 2 or 3, but then still claims to be building a
universal wheel.  As a result, pip may cache a wheel with the futures
dependency on Python 2 which will then be uninstallable on Python 3, or
conversely may cache a wheel without the futures dependency on Python 3
and then install it without futures on Python 2 as well.

For now, the simplest fix is to declare the dependency ourselves.  Since
the latest versions of gunicorn upstream have now dropped Python 2
support, it's unlikely that the gthread extra is going to grow new
dependencies on Python 2.