Fix clean build without cached wheels
Upgrade to setuptools 44.0.0 and wheel 0.34.2.  setuptools now uses pip
rather than easy_install to handle setup_requires; it still supports the
easy_install find_links option in .pydistutils.cfg, but no longer needs
or supports the allow_hosts option.

Install cffi earlier.  This is cumbersome, but it's needed because of
conflicts with the system-installed python-cffi-backend package in
system-site-packages mode; the recent pip upgrade caused us to trip over
this if we didn't have a cached wheel for PyNaCl.