Stop using --system-site-packages
We've been using "virtualenv --system-site-packages" to ensure that
Launchpad has access to some system-installed packages that can't or
shouldn't be installed as a normal Python dependency for various
reasons.  This strategy has always had some problems, but we've been
able to deal with them.

However, with recent versions of pip and Python 3, the situation gets
worse.  https://github.com/pypa/pip/issues/6264 describes the problem:
pip attempts to isolate itself from site-packages when building wheels,
leaving only its temporary one; but within a virtualenv created with
--system-site-packages it removes the virtualenv's site-packages but not
the system one (because virtualenv patches
distutils.sysconfig.get_python_lib).  It's not easy to see how pip might
reliably and portably escape the virtualenv in order to find the
system's site-packages directory.

(On Python 2 we get away with this by luck, because
distutils.sysconfig.get_python_lib returns
'.../env/lib/python2.7/site-packages' but
'.../env/local/lib/python2.7/site-packages' is on sys.path, and pip
doesn't realise that they're the same thing due to symlinks; we don't
get this stroke of luck on Python 3.)

Installing everything properly in the virtualenv is also difficult.
Several of the ones that remain don't exist as wheels or in another form
that pip can install.  We could fake this up (I went so far as to
prepare a script to convert the installed version of python-apt into a
wheel), but this would incur a significant maintenance burden.

Instead, symlink packages from the system site-packages directory into
the virtualenv in "make compile".  It's conceivable that this might
confuse pip slightly, but so far it doesn't seem to, and everything else
works fine.