Build a wheelhouse with relocation-safe paths
The paths to the wheels in pip's download cache (download-cache/wheels/)
depend on the hashes of the paths to the sdists from which they're
built.  As a result, a build on one machine that's pushed out to a
different file system path on another machine won't use the cached
wheels and will build them all again from scratch, which is very slow.

Instead, extend "make build_wheels" to use "pip wheel" to build a
wheelhouse of all our dependencies that forms part of the deployment
artifact and survives this kind of relocation.  This allows us to have
more reasonable deployment times again without having to commit lots of
wheels to lp-source-dependencies that might be architecture- or
series-dependent.