Fix librarian client socket handling for Python 3
On Python 3, only the 'r', 'w', and 'b' mode characters are supported.
'rwb' seems to be permitted to get a binary file object open for both
reading and writing, so use that.

We need a binary file object because we want it to be unbuffered; that
requires a few adjustments to the protocol implementation.