Remove XPI import support
As far as I can tell, this hasn't been used on production since about
2011, possibly due to changes in the Firefox release model.  It's
sufficiently complicated that it isn't worth keeping if it isn't being
used, although we can reintroduce it later if necessary.

The initial motivation for this was that it seems surprisingly difficult
to parse DTDs in modern Python without the non-Python-3-friendly copy of
parts of the old python-xml package that we've been keeping around in
sourcecode/old_xmlplus, but it turned into a larger opportunity to prune
unused code.

LP: #1016336