Add API to get branches/repositories ordered by modification date
This is useful to allow external source-code-archiving services to do so
efficiently.

For the most part, I've tried to ensure that the API looks similar
between branches and git_repositories; the exception is that
branches.getBranches doesn't currently take a target, since IHasBranches
isn't exported as a webservice entry type and since
IHasBranches.getBranches also exists.

Beware that the modification date is (of course) mutable, so client code
that tries to iterate over successive batches of the returned collection
may currently accidentally miss items between batches if
branches/repositories are modified between iterations.  It ought to be
possible for us to fix this by using StormRangeFactory, but in the
meantime, clients can work around this by calling
getBranches/getRepositories again with an appropriate
modified_since_date rather than traversing to the next batch of the
collection in the normal way.