Fix order of ZopeUtilityFixture cleanups
Fixture cleanups are run in reverse order of their addition, so it's
more correct to unregister the temporary utility before reregistering
the original.  This rarely matters in practice, but
XPIPOExporterTestCase.setUpTranslationImportQueueForTemplate managed to
trigger it: it temporarily reinstalls the ITranslationImporter utility
without its usual security proxy, so the temporary utility compares
equal to the original, and so reregistering the original first meant
that we ended up with no ITranslationImporter utility at all.