Parse Built-Using fields from uploaded binaries
These are parsed into a form that we'll later be able to use in the
dominator, with the goal of keeping source publications that are
referenced by Built-Using fields in active binary publications.

The most complicated bit is working out which source package releases a
given Built-Using field refers to, since there are edge cases where just
the name and version can be ambiguous.  We do the best we can by
following the archive dependencies of the build and finding sources that
it could plausibly have used.

Unresolvable entries in Built-Using, including ones that refer to
deleted source publications, will result in binary uploads being
rejected.

The grants for ArchiveDependency are needed because anything that
creates BinarySourceReference rows from package relationship fields
("source-package-name (= version)") needs to work out which SPR was
actually used by the build.  It's not clear how to do this completely
unambiguously, but we can derive a good approximation by trying the
build's archive dependencies in order.  To support this, anything that
creates BinarySourceReferences needs to be able to read
ArchiveDependency.

LP: #1868558