Don't make database queries in after-commit hooks
BaseRunnableJob's use of after-commit hooks was always a bit dodgy,
because the hook itself uses the database.  The semantics of this are
unclear at best and the transaction package doesn't document anything in
particular in this case; it's already bitten us once, requiring the
insertion of an extra transaction.abort() to stop the transaction begun
in runViaCelery from leaking into other code.  In fact, that should
probably have been a clue: the very fact that the after-commit hook
opened another transaction suggests the possibility of race conditions.

With transaction 3.0.0, the situation gets worse: trying to execute any
query in an after-commit hook raises an exception due to trying to join
a committed transaction.  It's possible this is in fact a bug, but this
approach has enough strikes against it by this point that it seems wiser
to avoid it altogether.  Instead, extract the relatively small amount of
state needed by runViaCelery in a before-commit hook, and block database
access in the corresponding after-commit hook.

Similarly, TestProcessAccepted.test_commits_after_each_item tried to
execute a database query in the afterCompletion hook of a synchronizer.
An analogous fix works there as well: move the database work to the
beforeCompletion hook.