Allow people with bug target roles to edit attachments
In particular, this ensures that bug supervisors can edit attachments on
bugs with corresponding tasks, which is needed by Ubuntu's retracers.

The implementation of this is borrowed from EditBug, and, as in that
case, isn't particularly fast for bugs with many tasks.  However,
checking whether we can edit a bug attachment isn't a hot operation (in
the UI, I don't believe we check this when viewing bugs, only on the
page to edit an attachment), so we can continue to get away with this
for the time being.

LP: #117752