Make TwistedJobRunner enter the virtualenv correctly
ampoule's normal child process startup mechanisms don't quite enter the
Launchpad virtualenv correctly.  I'd previously tried to deal with this
by adding `import _pythonpath` to its bootstrap code, but overlooked the
fact that it ran the system Python without the -S option, and so
retained the system site-packages directory in `sys.path`.  As of
zope.schema 6.0.0, this matters because the child process tries to
import zope.interface and finds the system one which is no longer
compatible enough with zope.schema in the virtualenv.

Instead, start the child processes in a way that's conceptually simpler
although it requires more code here: just run them with
`env/bin/python`.