Add test that sync-signingkeys commits changes
This mainly involved writing yet another thing that pretends to be a
signing service, because we needed something that could support use of
the signing service by subprocesses.  It's possible that we should also
refactor some of the existing tests to use this fake service to reduce
duplication (compare the test keyserver, for instance), though I haven't
attempted that here.