Port the rest of TestSlaveScannerScan to inlineCallbacks
We ported the most complicated ones in 2013, but this gets all the tests
in that class into roughly the same form.

Some tests used to assert that SlaveScanner.scan returned None.  This
hasn't been an interesting test since that method was converted to
inlineCallbacks in 2013: it now clearly unconditionally returns None.