Fix webhooks for team-owned private targets
In order to cope with webhook targets that use
DelegatedAuthorization-based security adapters, we had to switch to
using iter_authorization rather than looking up the security adapter
directly.  However, this breaks for private webhook targets that are
owned by a team, because teams don't have an account so we can't
construct a normal principal for them.  There's no appropriate principal
here, because we need to know about things that are visible to the whole
team in this case, not just to any particular member of the team.

However, iter_authorization only uses the principal to get hold of a
corresponding IPersonRoles, and most security adapters work just fine on
teams; so, to simplify all this and make it work in this case, we can
just pass an IPersonRoles directly.

LP: #1872740