Rewrite buildd-manager file fetching using subprocesses
This is similar to commit 8a6418c886433597e8aa113f53bd0b7dad397df2, but
using a process pool (via `ampoule`) rather than a thread pool.  The
threaded approach did a poor job of servicing established connections
when we deployed in on production, but it's possible that we were
running into trouble with Python's GIL, so let's see if subprocesses
will fare any better.

The new process pool is only used if the feature rule
"buildmaster.download_in_subprocess" is set, so we can easily switch
between the old and the new behaviour until we're confident.