Add IsTrue and IsFalse Storm extensions
I occasionally run into "IS TRUE" and "IS FALSE" when converting from
raw SQL strings, and the currently-available expressions aren't quite
synonymous when taking NULL into account.  This doesn't always matter in
practice because the column in question might be NOT NULL, but having an
exact synonym available in the query compiler means we can avoid having
to think so carefully about conversions.