Re-exec scripts with env/bin/python if necessary
Scripts have to have some kind of #! line, and with the current build
system it needs to be an absolute path (perhaps eventually this should
be refactored using entry_points).  This means we need to pick
/usr/bin/python2 or /usr/bin/python3, which makes it difficult to deal
with an incremental port to Python 3.

As an intermediate step, record the build-time major and minor version
of Python in env/python_version, and make _pythonpath.py (which all our
scripts import early) re-exec with env/bin/python if the current Python
version doesn't match that.