Avoid possible infinite loop in publishing.txt
If the list being shuffled to test that ArchiveSourcePublications
respects its ordering has length 0 or 1, then random.shuffle will
necessarily return the same ordering, so a loop that terminates only
when the ordering changes will never terminate.  This can only happen as
a result of an earlier failure, but an infinite loop is an unfriendly
way to report a test failure.

(This can be simplified when this test is rewritten as a unit test,
since unit tests don't barrel onward in the face of failures.)