Disentangle ArchiveSigningKeySet.{get,getSigningKey}
SigningUpload.findSigningHandlers wants to be able to get a SigningKey,
but for the other callers of ArchiveSigningKeySet.getSigningKey an
ArchiveSigningKey would be sufficient or better.  In particular,
SyncSigningKeysScript.inject currently has an inconsistent return type
and would be improved by being able to get an ArchiveSigningKey.

Split out ArchiveSigningKeySet.get, and re-express
ArchiveSigningKeySet.getSigningKey in terms of it.  Since an
ArchiveSigningKey uniquely identifies a SigningKey, no information is
lost.