Make LaunchpadConfig load ZConfig lazily
Loading ZConfig (launchpad.conf) is relatively heavyweight compared to
the rest of Launchpad's configuration system: it involves loading schema
resources from several packages and running an XML parser on them.
LaunchpadConfig only needs it for the "devmode" and "servers" settings,
which are used quite infrequently.  Load it lazily to save time.