Use six.python_2_unicode_compatible
Where appropriate, this makes it easier to define __str__ and
__unicode__ for Python 2 and only __str__ for Python 3.

In some places we define __unicode__ only to deliberately mark it as not
implemented.  However, those are related to URLs, where it seems to make
sense to define whatever the native __str__ is on each version, so mark
the __unicode__ definition as Python-2-only.