Move postprocessCandidate into build dispatch
BuildQueueSet.findBuildCandidates had a bug in its postprocessing logic:
it selected a prefix that was the *maximum* of twice the number of
builders being considered and 10, rather than the *minimum*, so ended up
doing an awful lot of unnecessary and often duplicated work on every
scan cycle.

However, it makes more sense to do the postprocessing once we actually
have an idle builder and are trying to dispatch something to it.  It's
unlikely that many candidates will be removed by postprocessing, and
even if that does happen then the worst case is that we spend a few
cycles dealing with superseding them all, which will self-resolve.  This
is better than spending more time every cycle doing database work.