Remove the bzr forking service
The performance optimisation idea was a reasonable one, but it never
quite made it to production
(https://bugs.launchpad.net/launchpad/+bug/732496), and now that we have
git hosting support it's unlikely that we're going to want to put much
effort into fixing a bzr optimisation.  It can always be resurrected
from history if need be.