Fix and simplify use of testtools.content
The get_bytes callable passed to Content is supposed to return an
iterable of bytes objects, not a single bytes object.  (On Python 2 we
get away with this, because iterating over a bytes object yields one
byte at a time; on Python 3 we don't, because iterating over a bytes
object yields the integer value of each byte.)

Pass buffer_now=False to some attach_file calls so that we get the
contents of the file at the end of the test, not the contents when the
test detail is added.

Use helper methods in testtools.content to simplify various longhand
constructions.