Show clean status duration on builders
Builders sometimes get stuck in the Cleaning status.  The
"manage-builders" tool in lp:ubuntu-archive-tools shows how long they've
been in that status, but the web UI doesn't, with the result that people
sometimes find it hard to tell the difference between
cleaning-because-recently-reset and cleaning-because-stuck.  Add an
indication to the web UI to make this more obvious.