Upgrade to python-openid2 3.2
This adds Python 3 support.

Our fork is no longer needed, because as of e6d54f6311 we force the
OpenID fetcher to Urllib2Fetcher in loggerhead as well as in
lp.services.webapp.login, and the fork was only ever needed to fix
CurlFetcher.

I've removed several patches for problems with python-openid that have
been fixed in python-openid2.