Fix editing team-owned OCI registry credentials
The permissions for Person:+edit-oci-registry-credentials are a bit
awkward for teams.  It looks as though it should use launchpad.Edit, but
for teams that's defined as requiring ownership of the team, while in
fact this is effectively a bulk editing interface for
OCIRegistryCredentials objects owned by the team in question and so
should only require membership of the team.  Work around this by
reducing the ZCML permissions to launchpad.View and doing additional
permission checking in the view.

In the process of testing this, I noticed that
PersonEditOCIRegistryCredentialsView.getEditFieldsRow sets up the owner
field with the wrong default (it needs to be given the Person object,
not its name), which led to OCIRegistryCredentials objects accidentally
changing owner when edited.  Fix this too.

LP: #1891078